+++ src/main/java/acme/features/technician/task/TechnicianTaskPublishService.java
@@ -1,4 +1,4 @@
-package acme.features.technician.tasks;
+package acme.features.technician.task;
 
 import org.springframework.beans.factory.annotation.Autowired;
