+++ src/main/java/acme/features/technician/task/TechnicianTaskUpdateService.java
@@ -1,4 +1,4 @@
-package acme.features.technician.tasks;
+package acme.features.technician.task;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -11,7 +11,7 @@ import acme.entities.tasks.TaskType;
 import acme.realms.technician.Technician;
 
 @GuiService
-public class TechnicianTaskCreateService extends AbstractGuiService<Technician, Task> {
+public class TechnicianTaskUpdateService extends AbstractGuiService<Technician, Task> {
 
 	// Internal state ---------------------------------------------------------
 
@@ -22,17 +22,29 @@ public class TechnicianTaskCreateService extends AbstractGuiService<Technician, 
 	// AbstractGuiService interface -------------------------------------------
 	@Override
 	public void authorise() {
-		super.getResponse().setAuthorised(true);
+		boolean exist;
+		Task task;
+		Technician technician;
+		int id;
+
+		id = super.getRequest().getData("id", int.class);
+		task = this.repository.findById(id);
+
+		exist = task != null;
+		if (exist) {
+			technician = (Technician) super.getRequest().getPrincipal().getActiveRealm();
+			if (technician.equals(task.getTechnician()))
+				super.getResponse().setAuthorised(true);
+		}
 	}
 
 	@Override
 	public void load() {
 		Task task;
-		Technician technician = (Technician) super.getRequest().getPrincipal().getActiveRealm();
+		int id;
 
-		task = new Task();
-		task.setDraftMode(true);
-		task.setTechnician(technician);
+		id = super.getRequest().getData("id", int.class);
+		task = this.repository.findById(id);
 
 		super.getBuffer().addData(task);
 	}
